# internal/interface/web/assets/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "go-time-recording API",
    "version": "1.0.0",
    "description": "Project time tracking.\n\nAuthenticate either with the session cookie obtained from `POST /auth/login`, or with a personal API token sent as `Authorization: Bearer <token>` or `X-API-Token: <token>`.\n\nA token carries no rights of its own: every request is evaluated against the owner's **current** role, so changing or revoking a role takes effect on the very next call.\n\nAll responses are wrapped by the framework as `{\"data\": ...}` on success and `{\"error\": {\"message\": ...}}` on failure."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiToken": []
    },
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "totp": {
                    "type": "string",
                    "description": "Required when the account has two-factor authentication enabled."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signed in; a session cookie is set. When the response says `totpRequired`, no cookie is set and the code must be supplied."
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Too many attempts"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A personal API token. Not subject to the CSRF check, since a browser never sends this header on its own."
      },
      "apiToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gtr_session",
        "description": "A browser session. State-changing calls made with this must also echo the gtr_csrf cookie in the X-CSRF-Token header and come from this origin; calls authenticated by a personal API token are exempt."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "The value of the gtr_csrf cookie, echoed back. Required alongside sessionCookie on POST, PUT, PATCH and DELETE."
      }
    },
...
  }
}